pkgrpArmy: Package for Army Shiny App - Group E Programming Tools in Data Science

PACKAGE: The package will recall two main function. The first one, 'Convert', return the graph and will create a list of points. The second one,'Orientation', is correlated to the first one as it will create a data-frame listing weapons class based on the sum of point from the list created before. Additionally the package contains the data-sets used and loaded directly in the functions mentioned before and that cand be loaded directly from the user. The first data-set, the 'points_army' data-set a data frame with 26 rows and 6 columns contianing several results from 5 type of exercises and the point gained from obtaining that resultb for each category. Example: Having a result between 2.80 and 2.85 in jumping will give you 24 points. The 'Performances' data-set, a data frame with 19 rows and 2 columns, is about the minimum points needd to reach a specific weapon class. Example: If after completing your exercises the sum of the results is 50, it means you can enter all the army class equals or below that requirement points, that in this case correspond to "Driver". SHINY: The package will also recall a shiny app regarding "Performance Scoreboard for Recruitment" with the 'militaryapp()' function. The app was built by group E and by inserting the result of some phisical exercises, will give you a possible weapon class of belonging. The app is entirely in English with video in French regarding the exercises to do.

Author

Maintainer: Jacopo Beghetto jacopo.beghetto@unil.ch

Authors:

Alistair Bisol alistair.bisol@unil.ch

Andrea Ferrazzo andrea.errazzo@unil.ch

Manoel Pidoux manoel.pidoux@unil.ch